fix for a nasty beach ball spinning when switching keyboard layouts in TotalFinder (http://getsatisfaction.com/binaryage/topics/mac_hangs_when_switching_language)
The problem lies in regenerateStringToKeyCodeMapping on first line where transformer is allocated. During the allocation, init is called and previously at that point kTISNotifySelectedKeyboardInputSourceChanged handler was registered. Normally is should not be a problem you say? Well, there are some complex keyboard layouts which somehow generate cascade of kTISNotifySelectedKeyboardInputSourceChanged notifications and system ends up with dozens of registered handlers waiting for next keyboard switch. I didn't investigate this to the core of the problem. But from my observation this affected all keyboard layout which have sub-layouts in System Preferences > Language & Text > Input Sources. For example Chinese, Kotoeri or Hangul.